Fix a flaky test case in DynamoHybridStore
An S3 bucket name has to be all lowercase.  Mostly randomAlphanumeric
creates names with >0 uppercase characters, so we get the "is not valid"
but occasionally it creates a valid bucket name, making this test flaky.

By always creating either a valid or invalid bucket name, this test
should no longer be flaky.