Check for the presence of the destination object first
We don't need to make a GetObject call against the source object if
there's no destination object to compare to.  In the case where we're
replicating something entirely fresh, this saves a lot of S3 calls.

In quick testing, it takes the time to replicate 250 objects with S3
from 2-3s to ~1.5s.